Fix handling of single-bidder scenarios in TPAL
This commit addresses a bug in the TPAL class related to handling scenarios with only one bidder. Previously, the code failed to correctly process bidder-specific misreports and bid profiles due to issues with array dimensions in such cases.

Furthermore `misr_bidder_i` method was modified: For multiple bidders, a boolean mask and `jnp.where` are used for simpler combination of original bids and misreports that avoids intermediate arrays.